server/requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
psycopg2==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2